requirements.txt
requests
beautifulsoup4
lxml
demoparser2
matplotlib
setuptools
patool